partition tests to appropriate modules
